rename: Header 공통 컴포넌트로 변경 - #58
